+++ Hattin/Types/Transposition.cs
@@ -3,7 +3,7 @@ namespace Hattin.Types
     public class Transposition
     {
         public GeneratedMove Move { get; set; }
-        public float Evaluation { get; set; }
+        public int Evaluation { get; set; }
         public int Depth { get; set; }
         public TranspositionEntryType Type { get; set; }
